rename group layer delegate
